+++ trunk/pol-core/pol/wrldsize.h
@@ -11,10 +11,10 @@ Notes
 #ifndef WRLDSIZE_H
 #define WRLDSIZE_H
 
-const unsigned WORLD_X = 6144;
-const unsigned WORLD_Y = 4096;
+const unsigned short WORLD_X = 6144;
+const unsigned short WORLD_Y = 4096;
 
-const int WORLD_MIN_Z  = -128;
-const int WORLD_MAX_Z  = +127;
+const short WORLD_MIN_Z  = -128;
+const short WORLD_MAX_Z  = +127;
 
 #endif
